+++ src/controller/getConfiguration.ts
@@ -9,21 +9,44 @@
 
 import vscode = require('vscode');
 
+export enum PrefixesAvailable {
+    hyphenMin = "-min",
+    dotMin = ".min",
+    minifiedHyphen = "minified-",
+    docMinified = ".minified"
+}
+
+export interface UserSettings {
+    hexDisabled: boolean;
+    disableHtml: boolean;
+    disableCss: boolean;
+    disableScss: boolean;
+    disableLess: boolean;
+    disableSass: boolean;
+    disableJson: boolean;
+    disableJsonc: boolean;
+    disableMessages: boolean;
+    minifyOnSave: boolean;
+    minifyOnSaveToNewFile: boolean;
+    prefix: PrefixesAvailable;
+    disableJavascript: boolean;
+    disableJavascriptReact: boolean;
+    disableTypescript: boolean;
+    openMinifiedDocument: boolean;
+}
+
 /**
  * Summary: The function getUserSettings gathers all settings from the user and returns them.
  *
  * @return {Object} with all the settings.
  */
-export default function getUserSettings() {
+export function getUserSettings(): UserSettings {
     const conf = vscode.workspace.getConfiguration('MinifyAll');
     if (conf.get('minifyOnSaveToNewFIle')) { // if the user is using a deprecated setting.
         vscode.window.showWarningMessage('You are using the deprecated setting "minifyOnSaveToNewFIle", please replace it with: "minifyOnSaveToNewFile" (mind the capital letter I)');
     }
     return {
         hexDisabled: conf.get('disableHexadecimalShortener'),
-        statusDisabled: conf.get('disableStatusbarInformation'),
-        priority: conf.get('statusbarPriority'),
-        alignment: conf.get('statusbarAlignment'),
         disableHtml: conf.get('disableHtml'),
         disableCss: conf.get('disableCss'),
         disableScss: conf.get('disableScss'),