+++ CHANGELOG.md
@@ -17,6 +17,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 * Extension recomendations in the .vscode/extension.json
 
+### Removed
+
+* The status bar information and all the settings that came with it.
+* Setting statusbarAlignment.
+* Setting disableStatusbarInformation.
+* Setting statusbarPriority.
+* Information about the new minified size.
+
 ## [**1.10.1**] - 2019-12-24
 
 ### Added